+++ packages/plugins/Attributions/src/index.ts
@@ -5,6 +5,9 @@ import { Attributions } from './components'
 import locales from './locales'
 import { makeStoreModule } from './store'
 
+import AttributionButton from './components/AttributionButton.vue'
+export { AttributionButton }
+
 export default (options: AttributionsConfiguration) => (instance: Vue) =>
   instance.$store.dispatch('addComponent', {
     name: 'attributions',