+++ packages/plugins/IconMenu/src/index.ts
@@ -4,6 +4,9 @@ import { IconMenu } from './components'
 import locales from './locales'
 import { makeStoreModule } from './store'
 
+import IconMenuButton from './components/IconMenuButton.vue'
+export { IconMenuButton }
+
 export default (options: IconMenuConfiguration) => (instance: Vue) =>
   instance.$store.dispatch('addComponent', {
     name: 'iconMenu',